Convert parameter_flash_storage
