+++ lib/CMakeLists.txt
@@ -95,3 +95,4 @@ add_subdirectory(cmp)
 add_subdirectory(cmp_mem_access)
 add_subdirectory(parameter)
 add_subdirectory(crc)
+add_subdirectory(parameter_flash_storage)
