--- lib/parameter/.gitignore
@@ -1,6 +1,0 @@
-CMakeLists.txt
-build
-dependencies
-
-# vim swap files
-*.sw[po]
