+++ lib/parameter/tests/parameter_test.cpp
@@ -3,30 +3,8 @@
 #include "parameter/parameter_port.h"
 #include <cstdlib>
 
-void parameter_port_lock(void)
+TEST_GROUP(ParameterNamespaceInit)
 {
-}
-
-void parameter_port_unlock(void)
-{
-}
-
-void parameter_port_assert(int condition)
-{
-    CHECK_TRUE(condition);
-}
-
-void* parameter_port_buffer_alloc(size_t size)
-{
-    return malloc(size);
-}
-
-void parameter_port_buffer_free(void* buffer)
-{
-    free(buffer);
-}
-
-TEST_GROUP (ParameterNamespaceInit) {
     parameter_namespace_t rootns;
     void setup(void)
     {
