+++ lib/parameter_flash_storage/parameter_flash_storage.c
@@ -1,7 +1,7 @@
 #include <string.h>
-#include "parameter_flash_storage.h"
-#include "parameter_flash_storage_private.h"
-#include "flash.h"
+#include "parameter_flash_storage/parameter_flash_storage.h"
+#include "parameter_flash_storage/parameter_flash_storage_private.h"
+#include "parameter_flash_storage/flash.h"
 #include <parameter/parameter_msgpack.h>
 #include <cmp/cmp.h>
 #include <cmp_mem_access/cmp_mem_access.h>
